ELEC-384: moved mock lights_blinker functions to lights_blinker.c so that build doesn't fail.
